Use database EntityType and add mixin.
