+++ app/schemas/emodel.py
@@ -1,9 +1,7 @@
 import uuid
-from typing import Literal
 
 from pydantic import BaseModel, ConfigDict
 
-from app.db.types import EntityType
 from app.schemas.annotation import ETypeClassRead, MTypeClassRead
 from app.schemas.asset import AssetsMixin
 from app.schemas.base import (
@@ -11,6 +9,7 @@ from app.schemas.base import (
     AuthorizationOptionalPublicMixin,
     BrainRegionRead,
     CreationMixin,
+    EntityTypeMixin,
     IdentifiableMixin,
     SpeciesRead,
     StrainRead,
@@ -40,7 +39,7 @@ class EModelCreate(EModelBase, AuthorizationOptionalPublicMixin):
     exemplar_morphology_id: uuid.UUID
 
 
-class EModelRead(EModelBase, CreationMixin, AssetsMixin, AuthorizationMixin):
+class EModelRead(EModelBase, CreationMixin, AuthorizationMixin, EntityTypeMixin, AssetsMixin):
     id: uuid.UUID
     species: SpeciesRead
     strain: StrainRead | None
@@ -49,7 +48,6 @@ class EModelRead(EModelBase, CreationMixin, AssetsMixin, AuthorizationMixin):
     mtypes: list[MTypeClassRead] | None
     etypes: list[ETypeClassRead] | None
     exemplar_morphology: ExemplarMorphology
-    type: Literal[EntityType.emodel] = EntityType.emodel
 
 
 class EModelReadExpanded(EModelRead):
