+++ tests/test_electrical_cell_recording.py
@@ -1,6 +1,7 @@
 import pytest
 
 from app.db.model import ElectricalCellRecording, ElectricalRecordingStimulus
+from app.db.types import EntityType
 
 from .utils import (
     PROJECT_ID,
@@ -128,6 +129,7 @@ def test_create_one(client, subject_id, license_id, brain_region_id, json_data):
     assert data["brain_region"]["id"] == brain_region_id
     assert data["license"]["id"] == str(license_id)
     assert data["authorized_project_id"] == str(PROJECT_ID)
+    assert data["type"] == EntityType.electrical_cell_recording
 
 
 def test_read_one(client, subject_id, license_id, brain_region_id, trace_id):
@@ -145,6 +147,7 @@ def test_read_one(client, subject_id, license_id, brain_region_id, trace_id):
     assert data["authorized_project_id"] == PROJECT_ID
     assert len(data["stimuli"]) == 2
     assert len(data["assets"]) == 1
+    assert data["type"] == EntityType.electrical_cell_recording
 
 
 def test_missing(client):
